Fix #147 - Use a separate temp directory
