+++ avrohugger-core/src/main/scala/format/abstractions/JavaTreehugger.scala
@@ -3,8 +3,11 @@ package format
 package abstractions
 
 import stores.ClassStore
-
 import org.apache.avro.Schema
+import org.apache.avro.compiler.specific.SpecificCompiler
+
+import java.io.{BufferedWriter, File, FileNotFoundException, FileWriter, IOException}
+import java.nio.file.{FileSystems, Files}
 
 /** Parent to all ouput format treehuggers
   * Note: Java required for generating Enums (Specific API requires Java enums)
@@ -17,5 +20,55 @@ trait JavaTreehugger {
     classStore: ClassStore,
     namespace: Option[String],
     schema: Schema): String
-  
+
+  def writeJavaTempFile(
+    schema: Schema,
+    outDir: File): Unit = {
+    // Uses Avro's SpecificCompiler, which only compiles from files, thus we
+    // write the schema to a temp file so we can compile a Java enum from it.
+    val tempSchemaFile = File.createTempFile(outDir + "/" + schema.getName, ".avsc")
+    tempSchemaFile.deleteOnExit()
+    val out = new BufferedWriter(new FileWriter(tempSchemaFile))
+    out.write(schema.toString)
+    out.close()
+
+    try {
+      SpecificCompiler.compileSchema(tempSchemaFile, outDir)
+    }
+    catch {
+      case ex: FileNotFoundException =>
+        sys.error("File not found:" + ex)
+      case ex: IOException =>
+        sys.error("There was a problem using the file: " + ex)
+    }
+  }
+
+  protected def createTempDir(prefix: String): File =
+    Files.createTempDirectory(
+      FileSystems.getDefault.getPath(System.getProperty("java.io.tmpdir")),
+      prefix
+    ).toFile
+
+  protected def deleteTempDirOnExit(tempDir: File): Unit = {
+    def getFiles(f: File): Set[File] = {
+      Option(f.listFiles)
+        .map(a => a.toSet)
+        .getOrElse(Set.empty)
+    }
+    def getRecursively(f: File): Set[File] = {
+      val files = getFiles(f)
+      val subDirectories = files.filter(path => path.isDirectory)
+      subDirectories.flatMap(getRecursively) ++ files + tempDir
+    }
+    def sortByDepth(f1: File, f2: File): Boolean = {
+      def countLevels(f: File) = f.getAbsolutePath.count(c => c == '/')
+      countLevels(f1) > countLevels(f2)
+    }
+    val filesToDelete = getRecursively(tempDir)
+    val sortedFilesToDelete = filesToDelete.toList.sortWith(sortByDepth)
+    sortedFilesToDelete.foreach(file => {
+      if (getFiles(file).isEmpty) file.deleteOnExit()
+    })
+  }
+
 }
